refactor(kimi-k3): extract SiTU/attn-res cores and router-weight ops into dedicated modules
Review feedback: move the optimized compute cores out of model.py and
experts.py into nemo_automodel/components/models/kimi_k3/situ.py and
nemo_automodel/components/moe/optimized_ops.py. Also add the requested
Args/Returns docstring to _attn_res_core. Pure move plus imports; no
functional change (unit battery unchanged: 38 tests pass).
